feat : sign-in 커밋 (#7)
* feat : 카카오 로그인 API

* feat : 네이버 로그인 API

* feat : 자동 로그인 API

+++ src/main/java/com/example/ged/utils/JwtService.java
@@ -44,7 +44,7 @@ public class JwtService {
      * @return int
      * @throws BaseException
      */
-    public int getUserIdx() throws BaseException {
+    public Long getUserIdx() throws BaseException {
         // 1. JWT 추출
         String accessToken = getJwt();
         if (accessToken == null || accessToken.length() == 0) {
@@ -60,6 +60,6 @@ public class JwtService {
             throw new BaseException(INVALID_JWT);
         }
         // 3. userIdx 추출
-        return claims.getBody().get("userIdx", Integer.class);
+        return claims.getBody().get("userIdx", Long.class);
     }
 }
